+++ internal/serve/httphandler/verify_receiver_registration_handler_test.go
@@ -152,14 +152,12 @@ func Test_VerifyReceiverRegistrationHandler_validate(t *testing.T) {
 
 	models, err := data.NewModels(dbConnectionPool)
 	require.NoError(t, err)
-	mockAnchorPlatformService := anchorplatform.AnchorPlatformAPIServiceMock{}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			handler := &VerifyReceiverRegistrationHandler{
-				Models:                   models,
-				AnchorPlatformAPIService: &mockAnchorPlatformService,
-				ReCAPTCHADisabled:        tc.isReCAPTCHADisabled,
+				Models:            models,
+				ReCAPTCHADisabled: tc.isReCAPTCHADisabled,
 			}
 
 			var requestBody io.Reader
@@ -211,12 +209,10 @@ func Test_VerifyReceiverRegistrationHandler_processReceiverVerificationPII(t *te
 	models, err := data.NewModels(dbConnectionPool)
 	require.NoError(t, err)
 
-	mockAnchorPlatformService := anchorplatform.AnchorPlatformAPIServiceMock{}
 	reCAPTCHAValidator := &validators.ReCAPTCHAValidatorMock{}
 	handler := &VerifyReceiverRegistrationHandler{
-		Models:                   models,
-		AnchorPlatformAPIService: &mockAnchorPlatformService,
-		ReCAPTCHAValidator:       reCAPTCHAValidator,
+		Models:             models,
+		ReCAPTCHAValidator: reCAPTCHAValidator,
 	}
 
 	// receiver without receiver_verification row:
@@ -371,10 +367,8 @@ func Test_VerifyReceiverRegistrationHandler_processReceiverWalletOTP(t *testing.
 
 	models, err := data.NewModels(dbConnectionPool)
 	require.NoError(t, err)
-	mockAnchorPlatformService := anchorplatform.AnchorPlatformAPIServiceMock{}
 	handler := &VerifyReceiverRegistrationHandler{
-		Models:                   models,
-		AnchorPlatformAPIService: &mockAnchorPlatformService,
+		Models: models,
 	}
 
 	// create valid sep24 token
@@ -498,91 +492,6 @@ func Test_VerifyReceiverRegistrationHandler_processReceiverWalletOTP(t *testing.
 					assert.ErrorContains(t, err, wantErrContain)
 				}
 				require.False(t, wasAlreadyRegistered)
-			}
-		})
-	}
-}
-
-func Test_VerifyReceiverRegistrationHandler_processAnchorPlatformID(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	dbConnectionPool, err := db.OpenDBConnectionPool(dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	ctx := context.Background()
-	models, err := data.NewModels(dbConnectionPool)
-	require.NoError(t, err)
-	handler := &VerifyReceiverRegistrationHandler{Models: models}
-
-	// creeate fixtures
-	phoneNumber := "+380445555555"
-	defer data.DeleteAllFixtures(t, ctx, dbConnectionPool)
-	wallet := data.CreateWalletFixture(t, ctx, dbConnectionPool, "testWallet", "https://home.page", "home.page", "wallet123://")
-	receiver := data.CreateReceiverFixture(t, ctx, dbConnectionPool, &data.Receiver{PhoneNumber: phoneNumber})
-	receiverWallet := data.CreateReceiverWalletFixture(t, ctx, dbConnectionPool, receiver.ID, wallet.ID, data.ReadyReceiversWalletStatus)
-
-	// create valid sep24 token
-	sep24Claims := &anchorplatform.SEP24JWTClaims{
-		ClientDomainClaim: wallet.SEP10ClientDomain,
-		RegisteredClaims: jwt.RegisteredClaims{
-			ID:        "test-transaction-id",
-			Subject:   "GBLTXF46JTCGMWFJASQLVXMMA36IPYTDCN4EN73HRXCGDCGYBZM3A444",
-			ExpiresAt: jwt.NewNumericDate(time.Now().Add(5 * time.Minute)),
-		},
-	}
-
-	// mocks
-	apTxPatch := anchorplatform.APSep24TransactionPatchPostRegistration{
-		ID:     "test-transaction-id",
-		Status: "pending_anchor",
-		SEP:    "24",
-	}
-
-	testCases := []struct {
-		name            string
-		mockReturnError error
-		wantErrContains string
-	}{
-		{
-			name:            "returns an error if the Anchor Platdorm API returns an error",
-			mockReturnError: fmt.Errorf("error updating transaction on anchor platform"),
-			wantErrContains: "error updating transaction on anchor platform",
-		},
-		{
-			name: "🎉 successfully updates the transaction on the Anchor Platform",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// dbTX
-			dbTx, err := dbConnectionPool.BeginTxx(ctx, nil)
-			require.NoError(t, err)
-			defer func() {
-				require.NoError(t, dbTx.Rollback())
-			}()
-
-			// mocks
-			mockAnchorPlatformService := anchorplatform.AnchorPlatformAPIServiceMock{}
-			defer mockAnchorPlatformService.AssertExpectations(t)
-			handler.AnchorPlatformAPIService = &mockAnchorPlatformService
-			mockAnchorPlatformService.
-				On("PatchAnchorTransactionsPostRegistration", mock.Anything, apTxPatch).
-				Return(tc.mockReturnError).Once()
-
-			// assertions
-			err = handler.processAnchorPlatformID(ctx, dbTx, *sep24Claims, *receiverWallet)
-			if tc.wantErrContains == "" {
-				require.NoError(t, err)
-
-				// make sure the receiverWallet was updated in the DB with the anchor platform transaction ID
-				var rw data.ReceiverWallet
-				err = dbTx.GetContext(ctx, &rw, "SELECT id, anchor_platform_transaction_id FROM receiver_wallets WHERE id = $1", receiverWallet.ID)
-				require.NoError(t, err)
-				assert.Equal(t, "test-transaction-id", rw.AnchorPlatformTransactionID)
-			} else {
-				require.ErrorContains(t, err, tc.wantErrContains)
 			}
 		})
 	}
@@ -1022,69 +931,6 @@ func Test_VerifyReceiverRegistrationHandler_VerifyReceiverRegistration(t *testin
 		require.Contains(t, buf.String(), wantErrContains)
 	})
 
-	t.Run("returns an error when processAnchorPlatformID() fails - anchor platform returns an error", func(t *testing.T) {
-		// mocks
-		reCAPTCHAValidator := &validators.ReCAPTCHAValidatorMock{}
-		defer reCAPTCHAValidator.AssertExpectations(t)
-		reCAPTCHAValidator.On("IsTokenValid", mock.Anything, "token").Return(true, nil).Once()
-
-		apTxPatch := anchorplatform.APSep24TransactionPatchPostRegistration{
-			ID:     "test-transaction-id",
-			Status: "pending_anchor",
-			SEP:    "24",
-		}
-		mockAnchorPlatformService := &anchorplatform.AnchorPlatformAPIServiceMock{}
-		defer mockAnchorPlatformService.AssertExpectations(t)
-		mockAnchorPlatformService.
-			On("PatchAnchorTransactionsPostRegistration", mock.Anything, apTxPatch).
-			Return(fmt.Errorf("error updating transaction on anchor platform")).Once()
-
-		// create handler
-		handler := &VerifyReceiverRegistrationHandler{
-			Models:                   models,
-			ReCAPTCHAValidator:       reCAPTCHAValidator,
-			AnchorPlatformAPIService: mockAnchorPlatformService,
-		}
-
-		// update database with the entries needed
-		defer data.DeleteAllReceiversFixtures(t, ctx, dbConnectionPool)
-		defer data.DeleteAllReceiverVerificationFixtures(t, ctx, dbConnectionPool)
-		defer data.DeleteAllReceiverWalletsFixtures(t, ctx, dbConnectionPool)
-		receiver := data.CreateReceiverFixture(t, ctx, dbConnectionPool, &data.Receiver{PhoneNumber: phoneNumber})
-		_ = data.CreateReceiverVerificationFixture(t, ctx, dbConnectionPool, data.ReceiverVerificationInsert{
-			ReceiverID:        receiver.ID,
-			VerificationField: data.VerificationTypeDateOfBirth,
-			VerificationValue: "1990-01-01",
-		})
-		_ = data.CreateReceiverWalletFixture(t, ctx, dbConnectionPool, receiver.ID, wallet.ID, data.ReadyReceiversWalletStatus)
-		_, err := models.ReceiverWallet.UpdateOTPByReceiverContactInfoAndWalletDomain(ctx, "+380445555555", wallet.SEP10ClientDomain, "123456")
-		require.NoError(t, err)
-
-		// set the logger to a buffer so we can check the error message
-		buf := new(strings.Builder)
-		log.DefaultLogger.SetOutput(buf)
-
-		// setup router and execute request
-		r.Post("/wallet-registration/verification", handler.VerifyReceiverRegistration)
-		req, err := http.NewRequest("POST", "/wallet-registration/verification", strings.NewReader(string(reqBody)))
-		require.NoError(t, err)
-		req = req.WithContext(context.WithValue(req.Context(), anchorplatform.SEP24ClaimsContextKey, validClaims))
-		rr := httptest.NewRecorder()
-		r.ServeHTTP(rr, req)
-
-		// execute and validate response
-		resp := rr.Result()
-		respBody, err := io.ReadAll(resp.Body)
-		require.NoError(t, err)
-		assert.Equal(t, http.StatusInternalServerError, resp.StatusCode)
-		wantBody := `{"error": "An internal error occurred while processing this request.", "error_code": "500_0"}`
-		assert.JSONEq(t, wantBody, string(respBody))
-
-		// validate logs
-		wantErrContains := fmt.Sprintf("processing anchor platform transaction ID: updating transaction with ID %s on anchor platform API", validClaims.TransactionID())
-		require.Contains(t, buf.String(), wantErrContains)
-	})
-
 	t.Run("🎉 successfully registers receiver's stellar address", func(t *testing.T) {
 		testCases := []struct {
 			name      string
@@ -1110,20 +956,10 @@ func Test_VerifyReceiverRegistrationHandler_VerifyReceiverRegistration(t *testin
 				defer reCAPTCHAValidator.AssertExpectations(t)
 				reCAPTCHAValidator.On("IsTokenValid", mock.Anything, "token").Return(true, nil).Once()
 
-				apTxPatch := anchorplatform.APSep24TransactionPatchPostRegistration{
-					ID:     "test-transaction-id",
-					Status: "pending_anchor",
-					SEP:    "24",
-				}
-				mockAnchorPlatformService := &anchorplatform.AnchorPlatformAPIServiceMock{}
-				defer mockAnchorPlatformService.AssertExpectations(t)
-				mockAnchorPlatformService.On("PatchAnchorTransactionsPostRegistration", mock.Anything, apTxPatch).Return(nil).Once()
-
 				// create handler
 				handler := &VerifyReceiverRegistrationHandler{
-					Models:                   models,
-					ReCAPTCHAValidator:       reCAPTCHAValidator,
-					AnchorPlatformAPIService: mockAnchorPlatformService,
+					Models:             models,
+					ReCAPTCHAValidator: reCAPTCHAValidator,
 				}
 
 				// update database with the entries needed
@@ -1215,20 +1051,10 @@ func Test_VerifyReceiverRegistrationHandler_VerifyReceiverRegistration(t *testin
 				defer reCAPTCHAValidator.AssertExpectations(t)
 				reCAPTCHAValidator.On("IsTokenValid", mock.Anything, "token").Return(true, nil)
 
-				apTxPatch := anchorplatform.APSep24TransactionPatchPostRegistration{
-					ID:     "test-transaction-id",
-					Status: "pending_anchor",
-					SEP:    "24",
-				}
-				mockAnchorPlatformService := &anchorplatform.AnchorPlatformAPIServiceMock{}
-				defer mockAnchorPlatformService.AssertExpectations(t)
-				mockAnchorPlatformService.On("PatchAnchorTransactionsPostRegistration", mock.Anything, apTxPatch).Return(nil)
-
 				// create handler
 				handler := &VerifyReceiverRegistrationHandler{
-					Models:                   models,
-					ReCAPTCHAValidator:       reCAPTCHAValidator,
-					AnchorPlatformAPIService: mockAnchorPlatformService,
+					Models:             models,
+					ReCAPTCHAValidator: reCAPTCHAValidator,
 				}
 
 				// update database with the entries needed
@@ -1384,15 +1210,6 @@ func Test_VerifyReceiverRegistrationHandler_VerifyReceiverRegistration(t *testin
 				defer reCAPTCHAValidator.AssertExpectations(t)
 				reCAPTCHAValidator.On("IsTokenValid", mock.Anything, "token").Return(true, nil).Once()
 
-				apTxPatch := anchorplatform.APSep24TransactionPatchPostRegistration{
-					ID:     "test-transaction-id",
-					Status: "pending_anchor",
-					SEP:    "24",
-				}
-				mockAnchorPlatformService := &anchorplatform.AnchorPlatformAPIServiceMock{}
-				defer mockAnchorPlatformService.AssertExpectations(t)
-				mockAnchorPlatformService.On("PatchAnchorTransactionsPostRegistration", mock.Anything, apTxPatch).Return(nil).Once()
-
 				mockCrashTracker := &crashtracker.MockCrashTrackerClient{}
 				defer mockCrashTracker.AssertExpectations(t)
 				mockEventProducer := events.NewMockProducer(t)
@@ -1434,7 +1251,6 @@ func Test_VerifyReceiverRegistrationHandler_VerifyReceiverRegistration(t *testin
 				handler := &VerifyReceiverRegistrationHandler{
 					Models:                      models,
 					ReCAPTCHAValidator:          reCAPTCHAValidator,
-					AnchorPlatformAPIService:    mockAnchorPlatformService,
 					EventProducer:               mockEventProducer,
 					CrashTrackerClient:          mockCrashTracker,
 					DistributionAccountResolver: distAccountResolverMock,